pyproject.toml
[project]
name = "demo-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.71.0",
    "arxiv>=2.2.0",
    "mcp>=1.18.0",
    "nest-asyncio>=1.6.0",
    "openai>=2.5.0",
    "python-dotenv>=1.1.1",
]
